fix #23: formula was wrong, and also organize files
