+++ src/engine/rays.c
@@ -14,8 +14,6 @@ t_ray	raycast(double u, double v, t_scene *scene)
 	cam = scene->camera;
 	cam_ray = camera_ray(cam, u, v);
 	world_hit = hit_objects(&cam_ray, scene);
-	/* after hit_objects, world_hit.object is the closest object hit by the ray
-	or NULL if no object was hit */
 	cam_ray.color = raycolor(cam_ray, &world_hit, scene);
 	return (cam_ray);
 }
@@ -46,20 +44,8 @@ t_color	raycolor(t_ray ray, t_hit *hit, t_scene *scene)
 	t_color		color;
 	double		t;
 
-//	if (hit->object && hit->object->type == OBJ_SPHERE)
-//		return (color_sphere((t_sphere *)hit->object->ptr, scene, hit));
-//	if (hit->object && hit->object->type == OBJ_PLANE)
-//		return (color_plane((t_plane *) hit->object->ptr, scene, hit));
-/*
-	if (hit->object && hit->object->type == OBJ_SPHERE)
-		return (phong_sphere((t_sphere *)hit->object->ptr, scene, hit));
-	if (hit->object && hit->object->type == OBJ_PLANE)
-		return (phong_plane((t_plane *) hit->object->ptr, scene, hit));
-
-*/	// Try same function for all objects
 	if (hit->t > EPSILON)
 		return (phong_light(scene, hit));
-	// sky color part
 	unit_direction = vec3_unit(ray.direction);
 	t = 0.5 * (unit_direction.y + 1.0);
 	c = vec3_sum(vec3_multk((t_v3){1.0, 1.0, 1.0}, 1.0 - t), \
@@ -76,3 +62,4 @@ t_v3	ray_at(t_ray *ray, double t)
 	r = vec3_sum(ray->origin, vec3_multk(ray->direction, t));
 	return (r);
 }
+
